backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.7.1
pydantic==2.5.0
python-multipart==0.0.6
PyPDF2==3.0.1
langchain==0.0.340
langchain-anthropic==0.1.1
supabase==2.0.3
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0